Applied cross-runtime build and test template.

+++ build/rollup.browser.config.js
@@ -1,7 +1,7 @@
 import path from 'path'
 import fs from 'fs'
-import config from './config.js'
-import NgnPlugin from './rollup-plugin-ngn.js'
+import config from './lib/config.js'
+import Build from './lib/build.js'
 import { terser } from 'rollup-plugin-terser'
 import babel from 'rollup-plugin-babel'
 
@@ -9,77 +9,76 @@ import babel from 'rollup-plugin-babel'
 import { install } from 'source-map-support'
 install()
 
-// Identify source file
-const input = path.resolve('../src/main.js')
+// Add our own functionality
+const build = new Build()
 
-// Add NGN rollup support
-const ngn = new NgnPlugin()
+// Identify source file
+const input = path.resolve(`../${build.pkg.main || '../src/index.js'}`)
 
 // Configure metadata for the build process.
 const rootdir = config.browserOutput // Main output directory
 let outdir = rootdir // Active output directory
 let configuration = [] // Rollup Configurations
 
+if (build.isUnnecessaryBuild(input, outdir)) {
+  process.exit(0)
+}
+
 // 1. Clean prior builds
 fs.rmdirSync(rootdir, { recursive: true })
 
 // Identify standard plugins
 const globalplugins = [
-  ngn.only('browser'),
-  ngn.applyVersion(ngn.version)
+  build.only('browser')
 ]
 
 // 2. Build Browser Production Package: Standard (Minified/Munged)
-outdir += '/browser-ngn'
-ngn.supportedBrowsers().forEach(edition => {
+outdir += `/browser-${build.name}`
+build.supportedBrowsers().forEach(edition => {
   console.log(`Generating ${edition} browser code.`)
   const plugins = globalplugins.slice()
 
   process.env.BROWSERSLIST_ENV = edition
   plugins.push(babel({
-    // presets: [['@babel/env', { targets: { node: true } }]],
+    presets: [['@babel/env']],
     plugins: [
       ['@babel/plugin-transform-flow-strip-types'],
       ['@babel/plugin-proposal-class-properties'/*, { 'loose': false } */],
       ['@babel/plugin-proposal-private-methods'/*, { 'loose': false } */]
     ]
-    // externalHelpersWhitelist: ['classPrivateFieldSet', 'classPrivateFieldGet']
   }))
 
-  plugins.push(terser({
-    module: true,
-    mangle: {
-      properties: true
-    },
-    compress: {
-      drop_console: true,
-      passes: 2,
-      warnings: true
-    }
-  }))
+  let terserCfg = config.terser
+  terserCfg.module = edition === 'current'
+
+  plugins.push(terser(terserCfg))
 
   configuration.push({
     input,
     plugins,
     output: {
-      // banner: config.banner,
-      file: `${outdir}/${ngn.name}-${ngn.version}${edition !== 'current' ? '-' + edition : ''}.min.js`,
+      banner: config.banner,
+      file: `${outdir}/${build.name}-${build.version}${edition !== 'current' ? '-' + edition : ''}.min.js`,
       format: edition === 'current' ? 'esm' : 'iife',
       sourcemap: true,
-      name: 'NGN' // namespace applied to window object
+      name: build.name // namespace applied to window object
     }
   })
 
   if (edition === 'current') {
+    plugins.pop()
+    terserCfg.module = false
+    terserCfg.compress.module = false
+    plugins.push(terser(terserCfg))
     configuration.push({
       input,
       plugins,
       output: {
-        // banner: config.banner,
-        file: `${outdir}/${ngn.name}-${ngn.version}-global.min.js`,
+        banner: config.banner,
+        file: `${outdir}/${build.name}-${build.version}-global.min.js`,
         format: 'iife',
         sourcemap: true,
-        name: 'NGN' // namespace applied to window object
+        name: build.name // namespace applied to window object
       }
     })
   }
